+++ ParkingGarageSimulator/src/nl/PriorIT/src/ParkingSimulator/core/Main.java
@@ -1,19 +1,35 @@
 package nl.PriorIT.src.ParkingSimulator.core;
 
+import java.awt.BorderLayout;
+import java.awt.Container;
+
 import nl.PriorIT.src.ParkingSimulator.logic.Model;
 import nl.PriorIT.src.ParkingSimulator.view.CarParkView;
+
 
 public class Main {
 
 	private Model simulatorModel;
-	private Model Model;
+	private CarParkView cpView;
 	//private CarParkView carParkView;
     /**
      * @param args
      */
+	public Main () {
+		simulatorModel = new Model(3, 6, 30);
+    	cpView = new CarParkView(simulatorModel);
+    	
+    	
+    	Container contentPane = getContentPane();
+        contentPane.add(carParkView, BorderLayout.CENTER);
+        pack();
+        setVisible(true);
+
+        carParkView.updateView();
+	}
 	
     public static void main(String[] args) {
-    	new Model(3, 6, 30);
+    	new Main();
 	}
