+++ ParkingGarageSimulator/src/nl/PriorIT/src/ParkingSimulator/view/CarParkView.java
@@ -13,11 +13,12 @@ public class CarParkView extends JPanel {
         
         private Dimension size;
         private Image carParkImage;
+        private Model model;
     
         /**
          * Constructor for objects of class CarPark
          */
-        public CarParkView() {
+        public CarParkView(Model model) {
             size = new Dimension(0, 0);
         }
     
@@ -57,6 +58,26 @@ public class CarParkView extends JPanel {
                     60 + location.getPlace() * 10,
                     20 - 1,
                     10 - 1); // TODO use dynamic size or constants
+        }
+        
+        public void updateView() {
+            // Create a new car park image if the size has changed.
+            if (!size.equals(getSize())) {
+                size = getSize();
+                carParkImage = this.createImage(size.width, size.height);
+            }
+            Graphics graphics = this.getGraphics();
+            for(int floor = 0; floor < model.getNumberOfFloors(); floor++) {
+                for(int row = 0; row < model.getNumberOfRows(); row++) {
+                    for(int place = 0; place < model.getNumberOfPlaces(); place++) {
+                        Location location = new Location(floor, row, place);
+                        Car car = model.getCarAt(location);
+                        Color color = car == null ? Color.white : car.getColor();
+                        this.drawPlace(graphics, location, color);
+                    }
+                }
+            }
+            this.repaint();
         }
     }
